% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialApprox-server.R
\name{ba_plots}
\alias{ba_plots}
\title{Create a plot of Binomial, Normal, and Poisson densities}
\usage{
ba_plots(n, p, plot_normal, plot_poisson)
}
\description{
Create a plot of Binomial, Normal, and Poisson densities
}
