% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialApprox-server.R
\name{ba_probs}
\alias{ba_probs}
\title{Compute probabilities for Binomial, Poisson, Normal, and correct Normal}
\usage{
ba_probs(n, p, a, b)
}
\description{
Compute probabilities for Binomial, Poisson, Normal, and correct Normal
}
