% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{binomialApproximation}
\alias{binomialApproximation}
\title{Compares binomial distribution to Normal and Poisson approximations}
\usage{
binomialApproximation()
}
\description{
Run a Shiny app in your web browser to compare the binomial distribution to
its approximation using the Normal and Poisson distributions.
}
